requirements.txt
appdirs==1.4.4
black==20.8b1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
google-api-core==1.24.1
google-api-python-client==1.8.0
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.16
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.5
pathspec==0.8.1
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
rsa==4.6
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
